deleting [action] route
